+++ Editor/MultiSceneToolsConfig_Editor.cs
@@ -3,13 +3,14 @@
 // *   Copyright (C) 2025 Henrik Hustoft
 // *
 // *   Check the Unity Asset Store for licensing information
+// *   https://assetstore.unity.com/packages/tools/utilities/multi-scene-tools-lite-304636
+// *   https://unity.com/legal/as-terms
 
 using UnityEngine;
 using UnityEditor;
 using HH.MultiSceneTools;
 using System.Reflection;
 using UnityEngine.SceneManagement;
-using UnityEditor.UIElements;
 
 namespace HH.MultiSceneToolsEditor
 {