+++ Editor/MultiSceneToolsEditorExtension.cs
@@ -3,8 +3,9 @@
 // *   Copyright (C) 2025 Henrik Hustoft
 // *
 // *   Check the Unity Asset Store for licensing information
+// *   https://assetstore.unity.com/packages/tools/utilities/multi-scene-tools-lite-304636
+// *   https://unity.com/legal/as-terms
 
-using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using HH.MultiSceneTools;