+++ Editor/MultiSceneToolsHierarchyStyle_Editor.cs
@@ -3,12 +3,13 @@
 // *   Copyright (C) 2025 Henrik Hustoft
 // *
 // *   Check the Unity Asset Store for licensing information
+// *   https://assetstore.unity.com/packages/tools/utilities/multi-scene-tools-lite-304636
+// *   https://unity.com/legal/as-terms
 
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEditor;
 using UnityEditor.SceneManagement;
-
 using HH.MultiSceneTools;
 
 namespace HH.MultiSceneToolsEditor