+++ Editor/MultiSceneToolsStartup.cs
@@ -3,6 +3,8 @@
 // *   Copyright (C) 2025 Henrik Hustoft
 // *
 // *   Check the Unity Asset Store for licensing information
+// *   https://assetstore.unity.com/packages/tools/utilities/multi-scene-tools-lite-304636
+// *   https://unity.com/legal/as-terms
 
 using UnityEditor;
 using HH.MultiSceneTools;
@@ -17,7 +19,7 @@ namespace HH.MultiSceneToolsEditor
         static void Startup()
         {
             // Events.registeredPackages += CheckUpdates;
-            
+
             CheckUpdates(true);
 
             if(!MultiSceneToolsConfig.instance)
@@ -34,6 +36,7 @@ namespace HH.MultiSceneToolsEditor
             if(MultiSceneToolsConfig.instance == null && tryOpenWizard)
             {
                 OpenWizard();
+                UnityEngine.Debug.Log("Multi Scene Tools: Opened Setup Wizard because no config was found.");
                 return;
             }
 
@@ -101,7 +104,7 @@ namespace HH.MultiSceneToolsEditor
         //     {
         //         OpenWizard();
         //     }
-        // }
+        // }   
 
         public static void HasShownUpdate()
         {
@@ -115,7 +118,6 @@ namespace HH.MultiSceneToolsEditor
         static void OpenWizard()
         {
             MultiSceneToolsSetup_Wizard.MenuEntryCall();
-            UnityEngine.Debug.Log("Multi Scene Tools: Opened Setup Wizard");
         }
     }
 }